fix: resolve compatibility issue with crypto.randomUUID when creating model provider
