+++ web/src/pages/ai-assistant/modules/EditForm.tsx
@@ -32,6 +32,7 @@ import { DeepThink } from "./DeepThink";
 import { formatESSearchResult } from "@/service/request/es";
 import ModelSelect from "./ModelSelect";
 import { ToolsConfig } from "./ToolsConfig";
+import { getUUID } from "@/utils/common";
 
 interface AssistantFormProps {
   initialValues: any;
@@ -481,14 +482,3 @@ export const SuggestedChat = ({ value = [], onChange }: any) => {
     </div>
   );
 };
-
-function getUUID() {
-  if (crypto && typeof crypto.randomUUID === "function") {
-    return crypto.randomUUID();
-  }
-  return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, (c) => {
-    const r = crypto.getRandomValues(new Uint8Array(1))[0] % 16;
-    const v = c === "x" ? r : (r & 0x3) | 0x8;
-    return v.toString(16);
-  });
-}
