+++ web/src/utils/common.ts
@@ -73,3 +73,15 @@ export function getKeys(obj: Record<string, any>, parentKeys: string[] = []): st
 
   return keys;
 }
+
+export function getUUID() {
+  if (crypto && typeof crypto.randomUUID === "function") {
+    return crypto.randomUUID();
+  }
+  return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, (c) => {
+    const r = crypto.getRandomValues(new Uint8Array(1))[0] % 16;
+    const v = c === "x" ? r : (r & 0x3) | 0x8;
+    return v.toString(16);
+  });
+}
+
